pass config error to frontend
